# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: true;
}

export interface ErrorEnvelope {
  error: string;
}

export interface DeleteSuccess {
  success: true;
}

export interface SearchRequest {
  /**
     * @minLength 1
     * @maxLength 500
     */
  query: string;
}

export interface SearchResultItem {
  url: string;
  name: string;
  categoryId: string;
  categoryLabel: string;
  relevance?: string | null;
}

export interface SearchResponse {
  results: SearchResultItem[];
  aiPowered: boolean;
  query: string;
}

export interface SaveToolRequest {
  toolUrl: string;
  toolName: string;
  categoryId: string;
  categoryLabel: string;
}

export interface SavedToolItem {
  id: number;
  toolUrl: string;
  toolName: string;
  categoryId: string;
  categoryLabel: string;
  savedAt: string;
}

export interface SavedToolsResponse {
  saved: SavedToolItem[];
}

export interface SearchHistoryItem {
  id: number;
  query: string;
  resultCount: number;
  aiPowered: boolean;
  searchedAt: string;
}

export interface SearchHistoryResponse {
  history: SearchHistoryItem[];
}

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};