# lib/api-zod/src/generated/types/savedToolItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavedToolItem {
  id: number;
  toolUrl: string;
  toolName: string;
  categoryId: string;
  categoryLabel: string;
  savedAt: Date;
}