# lib/api-zod/src/generated/types/savedToolsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SavedToolItem } from './savedToolItem';

export interface SavedToolsResponse {
  saved: SavedToolItem[];
}