# lib/api-zod/src/generated/types/searchHistoryItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchHistoryItem {
  id: number;
  query: string;
  resultCount: number;
  aiPowered: boolean;
  searchedAt: Date;
}