# lib/api-zod/src/generated/types/searchRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchRequest {
  /**
     * @minLength 1
     * @maxLength 500
     */
  query: string;
}