# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResultItem } from './searchResultItem';

export interface SearchResponse {
  results: SearchResultItem[];
  aiPowered: boolean;
  query: string;
}