# lib/api-zod/src/generated/types/searchResultItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResultItem {
  url: string;
  name: string;
  categoryId: string;
  categoryLabel: string;
  relevance?: string | null;
}